enter click, messageReader and clearMessageField functions to helpers file, importing all done on one line in footer.js
