+++ src/ChatroomComponents/Footer.js
@@ -1,21 +1,9 @@
 import React, { Component } from "react";
 import { handlePostClick } from '../helpers.js';
+import { enterClick, messageReader, clearMesssageField } from '../helpers.js';
 import { sendMessage } from '../client.js';
 
-const messageReader = _ => document.getElementById('text-field').value;
-const clearMessageField = _ => document.getElementById('text-field').value = '';
 
-const enterClick = document.addEventListener("keyup", function(event){
-  var clickedButton = document.getElementById("send-msg");
-  var userText = document.getElementById("text-field");
-  if (userText.value && event.keyCode === 13) {
-    clickedButton.click();
-    clickedButton.classList.add("active");
-    setTimeout(function() {
-      clickedButton.classList.remove("active");
-    }, 200);
-  }
-});
