+++ src/helpers.js
@@ -8,18 +8,38 @@ function findWebsocketAddress() {
     'ws://finalprojectwebsocketlivechat.herokuapp.com/'
     : 'http://localhost:8080/'
 }
+const messageReader = _ => document.getElementById('text-field').value;
+const clearMessageField = _ => document.getElementById('text-field').value = '';
 
 function handlePostClick(name) {
   var string = document.getElementById('text-field').value;
   if(string) {
-  sendMessage({ author: name, content: messageReader() });
-  clearMessageField();
+    sendMessage({ author: name, content: messageReader() });
+    clearMessageField();
   }
   document.getElementById("text-field").focus();
+}
+
+function enterClick() {
+  document.addEventListener("keyup", function(event){
+    var clickedButton = document.getElementById("send-msg");
+    var userText = document.getElementById('text-field');
+    if (userText.value && event.keyCode === 13) {
+      clickedButton.click();
+      clickedButton.classList.add("active");
+      setTimeout(function() {
+        clickedButton.classList.remove("active");
+      }, 200);
+    }
+  });
+
 }
 
 export {
   updateScroll,
   handlePostClick,
+  enterClick,
+  messageReader,
+  clearMessageField,
   findWebsocketAddress
 };
